requirements.txt
python-telegram-bot>=20.0
httpx>=0.25.0
python-dotenv>=1.0.0
google-api-python-client>=2.120.0
google-auth>=2.29.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
ruff>=0.3.0
